+++ src/test/java/com/codeborne/selenide/webdriver/EdgeDriverFactoryTest.java
@@ -51,6 +51,7 @@ class EdgeDriverFactoryTest {
 
     Map<String, Object> options = edgeOptions(edgeOptions);
     assertThat(args(options)).containsExactly(
+      "--remote-allow-origins=*",
       "--proxy-bypass-list=<-loopback>",
       "--disable-dev-shm-usage",
       "--no-sandbox",