+++ statics/src/test/java/integration/CustomWebdriverTest.java
@@ -1,25 +1,28 @@
 package integration;
 
 import com.codeborne.selenide.WebDriverRunner;
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxOptions;
+
+import java.io.File;
+import java.io.IOException;
 
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Condition.visible;
+import static com.codeborne.selenide.DownloadOptions.using;
+import static com.codeborne.selenide.FileDownloadMode.FOLDER;
+import static com.codeborne.selenide.Selectors.byText;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.using;
 import static com.codeborne.selenide.WebDriverRunner.closeWebDriver;
 import static com.codeborne.selenide.WebDriverRunner.isChrome;
 import static com.codeborne.selenide.WebDriverRunner.isFirefox;
 import static com.codeborne.selenide.WebDriverRunner.setWebDriver;
+import static com.codeborne.selenide.files.FileFilters.withExtension;
+import static java.time.Duration.ofSeconds;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
 
@@ -30,22 +33,15 @@ final class CustomWebdriverTest extends IntegrationTest {
   @BeforeAll
   static void setUpWebdrivers() {
     assumeThat(isChrome() || isFirefox()).isTrue();
-
     closeWebDriver();
-    if (isFirefox()) WebDriverManager.firefoxdriver().setup();
-    if (isChrome()) WebDriverManager.chromedriver().setup();
   }
 
   @BeforeEach
   void setUpTwoBrowsers() {
     closeWebDriver();
 
-    browser1 = isFirefox() ?
-      new FirefoxDriver(addSslErrorIgnoreCapabilities(addHeadless(new FirefoxOptions()))) :
-      new ChromeDriver(addSslErrorIgnoreCapabilities(addHeadless(new ChromeOptions())));
-    browser2 = isFirefox() ?
-      new FirefoxDriver(addSslErrorIgnoreCapabilities(addHeadless(new FirefoxOptions()))) :
-      new ChromeDriver(addSslErrorIgnoreCapabilities(addHeadless(new ChromeOptions())));
+    browser1 = isFirefox() ? openFirefox() : openChrome();
+    browser2 = isFirefox() ? openFirefox() : openChrome();
   }
 
   @Test
@@ -59,7 +55,7 @@ final class CustomWebdriverTest extends IntegrationTest {
     $("h1").shouldBe(visible).shouldHave(text("File upload form"));
 
     setWebDriver(browser1);
-    $("h1").shouldBe(visible).shouldHave(text("Page with selects"));;
+    $("h1").shouldBe(visible).shouldHave(text("Page with selects"));
   }
 
   @Test
@@ -109,6 +105,21 @@ final class CustomWebdriverTest extends IntegrationTest {
 
     assertThat(WebDriverRunner.hasWebDriverStarted()).isTrue();
     $("h1").shouldBe(visible).shouldHave(text("Some big divs"));
+  }
+
+  @Test
+  void canDownloadFilesAfterUsing() throws IOException {
+    openFile("page_with_uploads.html");
+    using(browser2, () -> {
+      openFile("page_with_selects_without_jquery.html");
+    });
+
+    File downloadedFile = $(byText("Download me")).download(
+      using(FOLDER).withTimeout(ofSeconds(2)).withFilter(withExtension("txt"))
+    );
+
+    assertThat(downloadedFile.getName()).matches("hello_world.*\\.txt");
+    assertThat(downloadedFile).content().isEqualToIgnoringNewLines("Hello, WinRar!");
   }
 
   @AfterEach